requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
Pillow==10.3.0
numpy==1.26.4
scipy==1.13.0
opencv-python-headless==4.9.0.80
librosa==0.10.2
scikit-learn==1.4.2
aiofiles==23.2.1
python-jose==3.3.0
httpx==0.27.0
pytest==8.2.0
pytest-asyncio==0.23.6
